Finish arbitrary catch distribution and MSY reprojection feature
To change the TAC catch distribution and obtain an updated MSY (and other ref vars) call the setRecommendedTACbyF() method of the MseFramework class. This can be done using either C++ or R methods and re-uses the normal projection code. It can also be done with cluster processing for better performance.