arbitrary CPUE series support
Add ability to use arbitrary CPUE time series in MP projections with normalisation as per Dales request. Add ability to run projection with the historic initialisation starting backdated a number of years (nbackupyears). This was added as a temporary means to bypass Stock Synthesis recruitment spikes in the final years and is included for its utility. In the default case of nbackupyears=0 it behaves as normal.